requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
babel==2.17.0
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.3.0
Django==5.1.6
django-appconf==1.1.0
django-cors-headers==4.7.0
django-rest-framework==0.1.0
django-select2==8.4.1
django-tailwind==4.0.1
djangorestframework==3.16.0
docxcompose==1.4.0
docxtpl==0.19.1
fastapi==0.115.12
Flask==3.1.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.5.2
lxml==5.3.1
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==25.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
pytailwindcss==0.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.20
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
XlsxWriter==3.2.3
